[REFACTOR/#260] 네비게이션 관련 패키지명 변경
- Nickname 및 TermsOfService 화면 내 네비게이션 함수명을 새로운 구조에 맞게 수정
